requirements.txt
# ---------------------------
# Core ML / ASR
# ---------------------------
torch>=2.0
torchaudio>=2.0
transformers>=4.35
datasets>=2.14
accelerate>=0.23
peft>=0.7

# ---------------------------
# Audio processing
# ---------------------------
librosa>=0.10
soundfile>=0.12
numpy>=1.23
scipy>=1.10

# ---------------------------
# Evaluation
# ---------------------------
jiwer>=3.0
evaluate>=0.4

# ---------------------------
# Visualization
# ---------------------------
matplotlib>=3.7
pandas>=1.5

# ---------------------------
# Optional (phonetic analysis)
# ---------------------------
pypinyin>=0.50

# ---------------------------
# Optional (lightweight ASR)
# ---------------------------
vosk>=0.3

# ---------------------------
# Utilities
# ---------------------------
tqdm>=4.65


pydub>=0.25
ffmpeg-python>=0.2.0

opencc-python-reimplemented

librosa
soundfile
pandas
opencc-python-reimplemented
jiwer
transformers

vosk

torch
torchaudio
transformers
datasets
peft
accelerate
soundfile
librosa
sentencepiece



pip uninstall -y transformers huggingface-hub
pip install "transformers==4.57.1" "huggingface-hub==0.36.2"